pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ados-drone-agent"
# DEC-107 Bug #24: version is sourced from src/ados/__init__.py per
# operating-rule 23 (single source of truth). Was hardcoded to 0.3.5
# while __init__.py sat at 0.3.9, causing pip metadata and `ados version`
# to disagree. Now both read from the same place.
dynamic = ["version"]
description = "ADOS Drone Agent — Software-defined drone network agent"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.11"
authors = [
    { name = "Altnautica", email = "team@altnautica.com" }
]
keywords = ["drone", "mavlink", "ardupilot", "companion-computer", "fleet-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "pymavlink>=2.4.41",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "paho-mqtt>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "websockets>=12.0",
    "structlog>=24.0.0",
    "pyserial>=3.5",
    "textual>=0.89.0",
    "psutil>=5.9.0",
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
    "zeroconf>=0.131.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "black>=24.1.0",
]
# DEC-112 / MSN-024: ground-station profile extras. Pulled only when the
# installer detects ground_station profile or when a developer runs
# `pip install -e ".[ground-station]"`. Drone profile never needs these.
ground-station = [
    "luma.oled>=3.13.0",
    "gpiozero>=2.0",
    "dbus-next>=0.2.3",
    "smbus2>=0.4.3",
    "evdev>=1.6.1",
]

[project.urls]
Homepage = "https://altnautica.com/drone-agent"
Repository = "https://github.com/altnautica/ADOSDroneAgent"
Documentation = "https://docs.altnautica.com/drone-agent"
Issues = "https://github.com/altnautica/ADOSDroneAgent/issues"

[project.scripts]
ados = "ados.cli.main:cli"
ados-agent = "ados.core.main:main"
ados-supervisor = "ados.core.supervisor:main"
ados-plugin-runner = "ados.plugins.runner:main"

[project.entry-points."ados.plugins"]
# Plugin entry points will be registered here

# MSN-028 Phase 4 Track A Wave 3: Peripheral Manager plugin group.
# External OEM partners contribute manifests via entry points or
# filesystem drop-in. No built-in plugins ship in Wave 3.
[project.entry-points."ados.peripherals"]

[tool.setuptools.dynamic]
version = {attr = "ados.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"ados.hal" = ["boards/*.yaml"]
"ados.core" = ["defaults.yaml"]
"ados.tui" = ["*.tcss"]


[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["UP017"]  # datetime.UTC requires Python 3.12+, we target 3.11

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
